+++ RELEASE-NOTES.md
@@ -23,7 +23,7 @@ The Uberenv project release numbers follow [Semantic Versioning](http://semver.o
 - Allow to prevent Uberenv from installing clingo setting `spack_setup_clingo`
   to `no` in `.uberenv_config.json`.
 - Adds the `--spack-debug` option to run spack spec/install commands in debug mode.
-- Adds `config:deprecated:true` to spack config, enabling deprecated package builds.
+- Adds the `--spack-allow-deprecated` option, to allow spack to build packages marked deprecated.
 
 ### Changed
 - All spack specs are now expressed inside single quotes to protect the parsing of complex flags.